+++ numerical/lp-problem/simplex-helpers/tableau.cpp
@@ -12,13 +12,13 @@ vector<vector<double>> create_tableau(
     const vector<double> &c,
     double z0)
 {
-  size_t m = A.size();
+  int m = A.size();
 
   // Copy the constraint coefficients into tableau
   vector<vector<double>> tableau = A;
 
   // Copy constraint constants into tableau
-  for (size_t i = 0; i < m; i++)
+  for (int i = 0; i < m; i++)
     tableau[i].push_back(b[i]);
 
   // Copy objective coefficients and constant into tableau
@@ -30,14 +30,14 @@ vector<vector<double>> create_tableau(
 
 /* THETA RATIO */
 
-double theta_ratio(const vector<double> &tableaurow, size_t pivotcol)
+double theta_ratio(const vector<double> &tableaurow, int pivotcol)
 {
   double entry;
 
   // Find the ratio only for a nonnegative entry
   if ((entry = tableaurow[pivotcol]) > 0)
   {
-    size_t n = tableaurow.size() - 1;
+    int n = tableaurow.size() - 1;
     return tableaurow[n] / entry;
   }
 
@@ -46,14 +46,14 @@ double theta_ratio(const vector<double> &tableaurow, size_t pivotcol)
 
 /* FIND PIVOT ROW */
 
-ptrdiff_t find_pivot_row(const vector<vector<double>> &tableau, size_t pivotcol)
+int find_pivot_row(const vector<vector<double>> &tableau, int pivotcol)
 {
-  size_t m = tableau.size() - 1;
-  size_t pivotrow;
+  int m = tableau.size() - 1;
+  int pivotrow;
 
   // Compute the theta ratio for each row and find the minimum
   double theta_min = numeric_limits<double>::infinity();
-  for (size_t i = 0; i < m; i++)
+  for (int i = 0; i < m; i++)
   {
     double theta = theta_ratio(tableau[i], pivotcol);
 
@@ -73,21 +73,19 @@ ptrdiff_t find_pivot_row(const vector<vector<double>> &tableau, size_t pivotcol)
 
 /* GET TABLEAU SOLUTION */
 
-vector<double> get_tableau_solution(
-    const vector<vector<double>> &tableau,
-    const vector<size_t> &basis)
+vector<double> get_tableau_solution(const vector<vector<double>> &tableau, const vector<int> &basis)
 {
-  size_t n = tableau[0].size() - 1;
+  int n = tableau[0].size() - 1;
   vector<double> solution(n);
 
   // Find the values of basic variables in the tableau
   // (nonbasic variables are set to 0)
-  for (size_t i = 0; i < n; i++)
+  for (int i = 0; i < n; i++)
   {
     auto it = find(basis.begin(), basis.end(), i);
     if (it != basis.end())
     {
-      size_t row = distance(basis.begin(), it);
+      int row = distance(basis.begin(), it);
       solution[i] = tableau[row][n];
     }
   }
