+++ numerical/lp-problem/simplex.cpp
@@ -11,19 +11,19 @@ pair<double, vector<double>> simplex(
     const vector<vector<double>> &A,
     const vector<double> &b,
     const vector<double> &c,
-    vector<size_t> &basis,
+    vector<int> &basis,
     double z0)
 {
   // Initial tableau.
   auto tableau = create_tableau(A, b, c, z0);
 
-  size_t m = A.size(), n = c.size();
-  ptrdiff_t pivotcol;
+  int m = A.size(), n = c.size();
+  int pivotcol;
 
   // Repeat as long as there are negative coefficients in the objective row
   while ((pivotcol = mostnegative(tableau[m].begin(), tableau[m].end() - 1)) != -1)
   {
-    ptrdiff_t pivotrow = find_pivot_row(tableau, pivotcol);
+    int pivotrow = find_pivot_row(tableau, pivotcol);
     if (pivotrow == -1)
     {
       // The problem is unbounded.
